LayoutTests/platform/ios/http/tests/events/device-orientation-motion-insecure-context-expected.txt

CONSOLE MESSAGE: Call to requestPermission() failed, reason: Browsing context is not secure.
CONSOLE MESSAGE: Call to requestPermission() failed, reason: Browsing context is not secure.
Tests that trying to set an event listener for deviceorientation and deviceorientation does not log an error in secure contexts.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Requesting device orientation access...
PASS result is "denied"

* Requesting device motion access...
PASS result is "denied"
PASS successfullyParsed is true

TEST COMPLETE
